# microservices/model-download/docs/user-guide/_assets/openapi.yaml
openapi: 3.0.0
info:
  title: Model Download Service API
  description: |
    API for downloading models from HuggingFace, Ollama, Ultralytics, Pipeline Zoo Models, Geti™ hubs and HLS, converting models to OpenVINO IR format.
    This service provides endpoints to download models and optionally convert them to 
    OpenVINO Intermediate Representation (IR) format for optimized inference on Intel hardware.

    **Features:**
    - Download models from multiple hubs (HuggingFace, Ollama, Ultralytics, Pipeline Zoo Models, Geti™, HLS)
    - Convert models to OpenVINO IR format with configurable precision and device targets for huggingface hub models
    - Asynchronous job processing with status tracking
    - Plugin-based architecture for extensibility

    **Note:** HUGGINGFACEHUB_API_TOKEN environment variable is optional and only required for downloading 
    gated models from HuggingFace or for model conversion. Public models can be downloaded without authentication.
  version: 1.1.0
  contact:
    name: Intel GenAI Team
servers:
  - url: http://localhost:8200/api/v1
    description: Local development server
  - url: http://{host}:{port}/api/v1
    description: Custom host deployment
    variables:
      host:
        default: localhost
        description: Server hostname or IP address
      port:
        default: 8200
        description: Server port
components:
  schemas:
...
    ModelDownloadRequest:
      type: object
      required:
        - models
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelRequest"
          description: List of models to download and/or convert
          minItems: 1
        parallel_downloads:
          type: boolean
          default: false
          description: Whether to download models in parallel (currently not implemented)
...
    ModelListResponse:
      type: object
      properties:
        hub:
          type: string
          description: The hub the models were listed from
          example: huggingface
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelListItem"
        total:
          type: integer
          nullable: true
          description: Total number of models available, when the hub reports it
        limit:
          type: integer
          description: Maximum number of items requested
        offset:
          type: integer
          description: Number of items skipped for pagination
paths:
...
  /hubs/{hub}/models:
    get:
      summary: List models available on a hub
      description: |
        Lists models available on a hub. Returns model names and details; the
        same model names can be passed as `name` to POST /api/v1/models/download.

        Listing is currently supported for the `huggingface` and `ultralytics`
        hubs. Hubs that do not support listing return `501`.
      operationId: listHubModels
      tags:
        - Models
      parameters:
        - in: path
          name: hub
          required: true
          schema:
            type: string
          description: The hub to list models from
          example: huggingface
        - in: query
          name: author
          required: false
          schema:
            type: string
          description: Filter by owner/organization (hub-specific)
          example: microsoft
        - in: query
          name: search
          required: false
          schema:
            type: string
          description: Substring search filter for model names
          example: phi
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
          description: Maximum number of items to return
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of items to skip for pagination
      responses:
        "200":
          description: Models listed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "400":
          description: Invalid pagination parameters, unknown hub, or hub not activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Unknown hub 'foo'"
        "401":
          description: Authentication failed for the hub
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Authentication required to list models"
        "501":
          description: Hub does not support listing models
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Hub 'ollama' does not support listing models"
        "502":
          description: Upstream hub request failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Failed to list models from hub 'huggingface'"

    post:
      summary: List models available on a hub using filters
      description: |
        Lists models available on a hub using a JSON request body for
        hub-specific filters and pagination. Listing is currently supported for
        the `huggingface` and `ultralytics` hubs. Hubs that do not support
        listing return `501`.
      operationId: listHubModelsWithBody
      tags:
        - Models
      parameters:
        - in: path
          name: hub
          required: true
          schema:
            type: string
          description: The hub to list models from
          example: huggingface
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelListRequest"
      responses:
        "200":
          description: Models listed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "400":
          description: Invalid pagination parameters, unknown hub, or hub not activated
        "401":
          description: Authentication failed for the hub
        "501":
          description: Hub does not support listing models
        "502":
          description: Upstream hub request failed
  /models/download:
    post:
      summary: Download and optionally convert models
      description: |
        Downloads one or more models from HuggingFace, Ollama, Ultralytics, Pipeline Zoo Models, hls or Geti™ and optionally converts them
        to OpenVINO IR format for huggingface hub models. This endpoint processes requests asynchronously and returns job IDs immediately.

        **Conversion Behavior:**
        Models will be converted to OpenVINO format if:
        1. `is_ovms` is set to `true` in the request
        2. `type` can be set to 'llm,embeddings,reranker, vlm or vision' in the request

        **Authentication:**
        - HUGGINGFACEHUB_API_TOKEN is optional for public HuggingFace models
        - HUGGINGFACEHUB_API_TOKEN is required for gated/private HuggingFace models and for conversion
        - No authentication needed for Ollama, Ultralytics, Pipeline Zoo Models, Geti™, or HLS models

        **Job Processing:**
        - Each model download creates a separate job
        - If conversion is requested, an additional conversion job is created
        - Use the returned job IDs to track progress via `/jobs/{job_id}` endpoint
      operationId: downloadModels
      parameters:
        - in: query
          name: download_path
          required: true
          schema:
            type: string
          description: Base path/subdirectory for model downloads (relative to MODELS_DIR)
          example: my-models
      tags:
        - Models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelDownloadRequest"
            examples:
              huggingface_download:
                summary: Download HuggingFace model
                value:
                  models:
                    - name: microsoft/Phi-3.5-mini-instruct
                      hub: huggingface
                      type: llm
                      is_ovms: false
              huggingface_with_conversion:
                summary: Download and convert to OpenVINO
                value:
                  models:
                    - name: microsoft/Phi-3.5-mini-instruct
                      hub: openvino
                      type: llm
                      is_ovms: true
                      config:
                        precision: int8
                        device: CPU
                        cache_size: 10
              ollama_download:
                summary: Download Ollama model
                value:
                  models:
                    - name: deepseek-r1
                      hub: ollama
              ultralytics_download:
                summary: Download Ultralytics model
                value:
                  models:
                    - name: yolov8s
                      hub: ultralytics
              pipeline_zoo_models_download:
                summary: Download Pipeline Zoo model
                value:
                  models:
                    - name: dbnet
                      hub: pipeline-zoo-models
              geti_download:
                summary: Download Geti™ model
                value:
                  models:
                    - name: yolox-tiny
                      hub: geti
              hls_download:
                summary: Download HLS model
                value:
                  models:
                    - name: ecg_8960_ir10_fp16
                      hub: hls
                      type: ai-ecg
              multiple_models:
                summary: Download multiple models
                value:
                  models:
                    - name: microsoft/Phi-3.5-mini-instruct
                      hub: huggingface
                      type: llm
                    - name: deepseek-r1
                      hub: ollama
                  parallel_downloads: true
      responses:
        "200":
          description: Request accepted and processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResponse"
              example:
                message: "Started processing 1 model(s)"
                job_ids: ["download_abc123", "convert_def456"]
                status: processing
        "400":
          description: Bad request - invalid hub, missing dependencies, or plugin not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
              examples:
                unsupported_hub:
                  summary: Unsupported model hub
                  value:
                    detail: "Unsupported model download/conversion detected. Supported methods are {'huggingface', 'ollama', 'ultralytics', 'pipeline-zoo-models', 'openvino', 'geti', 'hls'}."
                plugin_unavailable:
                  summary: Plugin dependencies not installed
                  value:
                    detail: "Plugin 'huggingface' is not available: Missing required dependencies: huggingface_hub"
        "422":
          description: Validation error in request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: array
                    items:
                      type: object
                      properties:
                        loc:
                          type: array
                          items:
                            type: string
                        msg:
                          type: string
                        type:
                          type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /jobs/{job_id}:
    get:
      summary: Get job status
      description: Retrieve the status and details of a specific job by its ID
      operationId: getJobStatus
      tags:
        - Jobs
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
          description: The unique job identifier
          example: download_abc123
      responses:
        "200":
          description: Job details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Job download_abc123 not found"
...
  /models/jobs:
    get:
      summary: Get jobs for a specific model
      description: Retrieve all jobs (downloads and conversions) related to a specific model
      operationId: getModelJobs
      tags:
        - Jobs
      parameters:
        - in: query
          name: model_name
          required: true
          schema:
            type: string
          description: The model name
          example: microsoft/Phi-3.5-mini-instruct
      responses:
        "200":
          description: Jobs for the model retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
        "404":
          description: No jobs found for the model
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "No jobs found for model microsoft/Phi-3.5-mini-instruct"
...